requirements.txt
requests>=2
pyttsx3>=2
openai>=1
PyPDF2>=3
PyYAML>=6
